requirements.txt
click==8.0.1
dlib==19.22.1
face-recognition==1.3.0
face-recognition-models==0.3.0
numpy==1.21.2
Pillow==8.3.1
